backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Plaid API integration
plaid-python==12.0.0

# ML and data science
scikit-learn==1.3.2
prophet==1.1.5
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4

# Task queue and scheduling
celery==5.3.4
redis==5.0.1

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# CORS and security
python-cors==1.0.0
cryptography==41.0.8

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Date and time utilities
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0 